Refactor cmd packages as minimal wrappers around component logical Main()
